pyproject.toml
[project]
name = "slack-bot-server"
version = "0.1.0"
description = ""
authors = [
    {name = "Theo Herveux",email = "theo.herveux@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "requests (>=2.32.4,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]

[tool.poetry]
packages = [{include = "slack_bot_server", from = "src"}]

[tool.poetry.requires-plugins]
poetry-plugin-export = ">=1.8"


[tool.poetry.group.dev.dependencies]
pre-commit = "^4.2.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
